feat: modify CartService.clearCart() to check if the cart exists first for CartServiceTest.shouldBeAbleToClearCart()
Refs: #6

+++ week1/src/main/java/com/example/week1/cart/CartService.java
@@ -28,6 +28,8 @@ public class CartService {
     }
 
     public void clearCart(String username) {
-        cartRepository.deleteById(username);
+        if(cartRepository.existsById(username)) {
+            cartRepository.deleteById(username);
+        }
     }
 }
